+++ ice-20201109/core/Models/SubYikeUserCreditResponseBody.cs
@@ -1,0 +1,38 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.ICE20201109.Models
+{
+    public class SubYikeUserCreditResponseBody : TeaModel {
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>NOT_ENOUGH_ALLOCATE_CREDIT_QUOTA</para>
+        /// </summary>
+        [NameInMap("ErrorCode")]
+        [Validation(Required=false)]
+        public string ErrorCode { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para><b><b>63E8B7C7-4812-46AD-0FA56029AC86</b></b></para>
+        /// </summary>
+        [NameInMap("RequestId")]
+        [Validation(Required=false)]
+        public string RequestId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>true</para>
+        /// </summary>
+        [NameInMap("Result")]
+        [Validation(Required=false)]
+        public bool? Result { get; set; }
+
+    }
+
+}